PaletteBlockCtrl must be aware of the current PaletteBlockSize
It is important to initialize the blocks with the same size

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/utils/XcosConstants.java
@@ -79,7 +79,7 @@ public final class XcosConstants extends ScilabGraphConstants {
         /**
          * @return block width
          */
-        public float getBlockWidth() {
+        public int getBlockWidth() {
             switch (this) {
                 case SMALL:
                     return 70;
@@ -95,7 +95,7 @@ public final class XcosConstants extends ScilabGraphConstants {
         /**
          * @return block height
          */
-        public float getBlockHeight() {
+        public int getBlockHeight() {
             switch (this) {
                 case SMALL:
                     return 70;
